+++ scripts/pull-upstream.sh
@@ -1,0 +1,16 @@
+#!/bin/bash
+# This script pulls the latest changes from the upstream repository's main branch
+# and merges them into the current branch.
+
+set -e
+
+UPSTREAM_REPO_PATH="../melanie/packages/rmm-schema"
+
+rsync -av --delete "$UPSTREAM_REPO_PATH/src/." "src/."
+rsync -av --delete \
+	"$UPSTREAM_REPO_PATH/CHANGELOG.md" \
+	"$UPSTREAM_REPO_PATH/package.json" \
+	"$UPSTREAM_REPO_PATH/README.md" \
+	"$UPSTREAM_REPO_PATH/ROADMAP.md" \
+	"$UPSTREAM_REPO_PATH/tsconfig.json" \
+	"."